playground/requirements.txt
opencopilot-ai
langchain
atlassian-python-api
